#178 Break the pack algorithm out into a node module
